+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+Performance optimization
+* position now takes a stream to rebuild as little as possible
+* layout builder replaced by a cached version to avoid rebuilding too often
+
 ## 0.1.0
 
 * First release